+++ packages/ckeditor5-ui/src/toolbar/toolbarview.js
@@ -16,6 +16,7 @@ import ToolbarLineBreakView from './toolbarlinebreakview';
 import ResizeObserver from '@ckeditor/ckeditor5-utils/src/dom/resizeobserver';
 import preventDefault from '../bindings/preventdefault.js';
 import Rect from '@ckeditor/ckeditor5-utils/src/dom/rect';
+import isVisible from '@ckeditor/ckeditor5-utils/src/dom/isvisible';
 import global from '@ckeditor/ckeditor5-utils/src/dom/global';
 import { createDropdown, addToolbarToDropdown } from '../dropdown/utils';
 import { logWarning } from '@ckeditor/ckeditor5-utils/src/ckeditorerror';
@@ -758,7 +759,7 @@ class DynamicGrouping {
 		// the toolbar is visible (the next ResizeObserver callback execution). This is handy because
 		// the grouping could be caused by increasing the #maxWidth when the toolbar was invisible and the next
 		// time it shows up, some items could actually be ungrouped (https://github.com/ckeditor/ckeditor5/issues/6575).
-		if ( !this.viewElement.offsetParent ) {
+		if ( !isVisible( this.viewElement ) ) {
 			this.shouldUpdateGroupingOnNextResize = true;
 
 			return;
