+++ packages/ckeditor5-ui/theme/components/input/input.css
