+++ packages/ckeditor5-utils/src/index.js
@@ -26,6 +26,7 @@ export { default as Rect } from './dom/rect';
 export { default as ResizeObserver } from './dom/resizeobserver';
 export { default as setDataInElement } from './dom/setdatainelement';
 export { default as toUnit } from './dom/tounit';
+export { default as isVisible } from './dom/isvisible';
 
 export * from './keyboard';
 export * from './language';
